+++ R/graphs.R
@@ -313,7 +313,8 @@ rebound_graphs_helper <- function(.path_data,
   g +  
     # Use the colour, linewidth, linetype, and shape columns/data directly.
     ggplot2::scale_colour_identity() + 
-    ggplot2::scale_linewidth(trans = "identity", guide = "none", range = c(linewidth_min, linewidth_max)) + 
+    # ggplot2::scale_linewidth(trans = "identity", guide = "none", range = c(linewidth_min, linewidth_max)) + 
+    ggplot2::scale_discrete_identity(aesthetics = "linewidth") +
     ggplot2::scale_size_identity() + 
     ggplot2::scale_linetype_identity() + 
     ggplot2::scale_shape_identity() + 
@@ -487,7 +488,8 @@ sensitivity_graphs <- function(.parametric_data = parametric_analysis(rebound_da
       ggplot2::geom_path(data = line_data,
                          mapping = ggplot2::aes(x = .data[[x_var]],
                                                 y = .data[[y_vals_col]],
-                                                size = .data[[line_var]],
+                                                # size = .data[[line_var]],
+                                                linewidth = .data[[line_var]],
                                                 linetype = .data[[line_var]],
                                                 colour = .data[[line_var]]), 
                          lineend = graph_params$lineend, 
@@ -512,7 +514,8 @@ sensitivity_graphs <- function(.parametric_data = parametric_analysis(rebound_da
       ggplot2::geom_path(data = line_data,
                          mapping = ggplot2::aes(x = .data[[x_var]],
                                                 y = .data[[y_vals_col]],
-                                                size = .data[[line_var]],
+                                                # size = .data[[line_var]],
+                                                linewidth = .data[[line_var]],
                                                 linetype = .data[[line_var]],
                                                 colour = .data[[line_var]]), 
                          lineend = graph_params$lineend, 
@@ -638,23 +641,42 @@ rebound_terms_graph <- function(.parametric_data = parametric_analysis(rebound_d
                                               Re_tot = graph_params$tot_linetype), 
                                    labels = legend_labs,
                                    breaks = Re_terms) + 
-    ggplot2::scale_size_manual(values = c(Re_dempl = graph_params$dempl_linewidth,
-                                          Re_emb = graph_params$emb_linewidth,
-                                          Re_md = graph_params$md_linewidth,
-                                          Re_cap = graph_params$cap_linewidth,
-                                          Re_empl = graph_params$empl_linewidth,
-                                          Re_dsub = graph_params$dsub_linewidth,
-                                          Re_isub = graph_params$isub_linewidth,
-                                          Re_sub = graph_params$sub_linewidth,
-                                          Re_dinc = graph_params$dinc_linewidth,
-                                          Re_iinc = graph_params$iinc_linewidth,
-                                          Re_inc = graph_params$inc_linewidth,
-                                          Re_micro = graph_params$micro_linewidth,
-                                          Re_macro = graph_params$macro_linewidth,
-                                          Re_dir = graph_params$dir_linewidth,
-                                          Re_indir = graph_params$indir_linewidth,
-                                          Re_tot = graph_params$tot_linewidth),
-                               labels = legend_labs,
-                               breaks = Re_terms)
+    # ggplot2::scale_size_manual(values = c(Re_dempl = graph_params$dempl_linewidth,
+    #                                       Re_emb = graph_params$emb_linewidth,
+    #                                       Re_md = graph_params$md_linewidth,
+    #                                       Re_cap = graph_params$cap_linewidth,
+    #                                       Re_empl = graph_params$empl_linewidth,
+    #                                       Re_dsub = graph_params$dsub_linewidth,
+    #                                       Re_isub = graph_params$isub_linewidth,
+    #                                       Re_sub = graph_params$sub_linewidth,
+    #                                       Re_dinc = graph_params$dinc_linewidth,
+    #                                       Re_iinc = graph_params$iinc_linewidth,
+    #                                       Re_inc = graph_params$inc_linewidth,
+    #                                       Re_micro = graph_params$micro_linewidth,
+    #                                       Re_macro = graph_params$macro_linewidth,
+    #                                       Re_dir = graph_params$dir_linewidth,
+    #                                       Re_indir = graph_params$indir_linewidth,
+    #                                       Re_tot = graph_params$tot_linewidth),
+    #                            labels = legend_labs,
+    #                            breaks = Re_terms)
+  ggplot2::scale_discrete_manual(aesthetics = "linewidth", 
+                                 values = c(Re_dempl = graph_params$dempl_linewidth,
+                                            Re_emb = graph_params$emb_linewidth,
+                                            Re_md = graph_params$md_linewidth,
+                                            Re_cap = graph_params$cap_linewidth,
+                                            Re_empl = graph_params$empl_linewidth,
+                                            Re_dsub = graph_params$dsub_linewidth,
+                                            Re_isub = graph_params$isub_linewidth,
+                                            Re_sub = graph_params$sub_linewidth,
+                                            Re_dinc = graph_params$dinc_linewidth,
+                                            Re_iinc = graph_params$iinc_linewidth,
+                                            Re_inc = graph_params$inc_linewidth,
+                                            Re_micro = graph_params$micro_linewidth,
+                                            Re_macro = graph_params$macro_linewidth,
+                                            Re_dir = graph_params$dir_linewidth,
+                                            Re_indir = graph_params$indir_linewidth,
+                                            Re_tot = graph_params$tot_linewidth),
+                                 labels = legend_labs,
+                                 breaks = Re_terms)
 }