+++ NEWS.md
@@ -1,3 +1,10 @@
+
+* Now using `ggplot2::scale_discrete_manual()` and 
+  `ggplot2::scale_discrete_identity()` for 
+  various rebound graphs, thereby 
+  eliminating some warnings.
+
+
 # ReboundTools 0.1.33 (2022-11-20) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.7340040.svg)](https://doi.org/10.5281/zenodo.7340040)
 
 * Now using the `linewidth` aesthetic in `ggplot2` where possible.